Append files
I had many .txt files as output of one of my projects. Then I had to re run some of then that were not saved on the first time. The second run files had the same name but new content that should be append to the first run files. 

This script was created to do the hard work and it is not a general purpose script.